Add Docker support and dependencies for Django app
Created Dockerfile and compose.yaml for running Dockerized Django app. Also added the necessary dependencies in the requirements.txt file. The Dockerfile is configured to create a non-privileged user to run the app and leverage Docker's caching system for dependencies, while the compose.yaml is set to run the server with PostgreSQL as the database.